Section removal complete
