+++ templates/sectionCreation/removeSection.html
@@ -7,52 +7,51 @@
 </head>
 <div class="col-auto text-center">
     <h1>TA Scheduler</h1>
-    <h2><small>Remove User</small></h2>
+    <h2><small>Remove Course</small></h2>
+</div>
 <body>
 <div class="row justify-content-center" >
     <div class="col-auto">
         <table class = "table table-responsive">
             <thead>
                 <tr>
-                    <th>ID Number</th>
-                    <th>Name</th>
-                    <th>Address</th>
-                    <th>Email</th>
-                    <th>Phone</th>
-                    <th>Role</th>
+                    <th>Section Code</th>
+                    <th>Course Number</th>
+                    <th>TA</th>
 
                 </tr>
             </thead>
 
+
             <tbody>
-                {% for row in users %}
+                {% for row in sections %}
                     <tr>
-                        <td>{{ row.IDNumber }}</td>
-                        <td>{{ row.name }}</td>
-                        <td>{{ row.address }}</td>
-                        <td>{{ row.email }}</td>
-                        <td>{{ row.phoneNumber }}</td>
-                        <td>{{ row.role }}</td>
+                        <td>{{ row.sectionCode }}</td>
+                        <td>{{ row.parentCode.courseNumber }}</td>
+                        <td>{{ row.TA.name }}</td>
                     </tr>
 
                 {% endfor %}
 
             </tbody>
+
+
+
         </table>
     </div>
 </div>
 <br>
     <div class="text-center" class="center-block">
-     <form class="text-center" style="text-align: center" action="/removeUser/" method="post">
+     <form class="text-center" style="text-align: center" action="/removeSection/" method="post">
          {% csrf_token %}
          <div class="form-group">
-                              <label value="t2" for="InputUser"><select id = "InputUser" name = "InputUser" class="form-control" required>
-                                              {% for row in users %}
-                                                <option>{{ row.IDNumber }} </option>
+                              <label value="t2" for="InputCourse"><select id = "InputCourse" name = "InputCourse" class="form-control" required>
+                                              {% for row in sections %}
+                                                <option>{{ row.sectionCode }} </option>
                                               {% endfor %}
                                             </select> </label>
-                    <button type="submit" class="btn btn-primary my-2 my-sm-0">Remove User</button>
-                    <a href="/users">
+                    <button type="submit" class="btn btn-primary my-2 my-sm-0">Remove Section</button>
+                    <a href="/sections">
                         Back
                     </a>
         </div>
@@ -60,7 +59,7 @@
     <br>
     {{ message }}
     </div>
-</div>
+
 
 </body>
 </html>
